feat: layout style size
