+++ src/Pages/Guide/UserGuideline.js
@@ -1,10 +1,9 @@
 import React from 'react'
-import guideTest from '../../IMG/guide3.png'
 
 const UserGuideline = () => {
     return (
         <div>
-            <img src={guideTest} alt="background" style={{ marginLeft: '350px' }} />
+            {/*<img src={guideTest} alt="background" style={{ marginLeft: '350px' }} />*/}
         </div>
     )
 }
